isolate post-db startup procedures
